Improvements to UI service with full screen message (#493)
* Removing the handler

* Simplyfying UI service

* Changes based on comments

* Adding global listener for any message

* Updating the global listeners

* Adding Current message in UIService

* Adding currentMessage in protocol

* Updating comments

* Updating tests

* changes

* Updating message monitor

* Simplying the ui services fullscreen message

* Updating the test

* Updating comments

* Updates with alert

* Updating lint errors

* Updating show method of fullscreen message

* Updating some tests

* Updating based on review and comments tests

* Updating listeners with message object and updating tests

* Fixing the Lint errors

* Changes for service

* Adding protocol to msg monitor service

* Adding a protocol UIMessaging for messages

* Updates to tests

* Fixing lint

* Updating tests

* Renaming and updating comments

* Remove unecessary code
