+++ AEPServices/Sources/ui/fullscreen/FullscreenMessage.swift
@@ -1,5 +1,5 @@
 /*
- Copyright 2020 Adobe. All rights reserved.
+ Copyright 2021 Adobe. All rights reserved.
  This file is licensed to you under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License. You may obtain a copy
  of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -11,43 +11,213 @@
  */
 
 import Foundation
+import UIKit
+import WebKit
 
-public class FullScreenMessage {
-    var fullScreenMessageHandler: FullScreenUIHandler?
+/// This class is used to create and display fullscreen messages on the currentt view.
+@objc(AEPFullscreenMessage)
+public class FullscreenMessage: NSObject, WKNavigationDelegate, UIMessaging {
+    private let LOG_PREFIX = "FullscreenMessage"
+    private let DOWNLOAD_CACHE = "adbdownloadcache"
+    private let HTML_EXTENSION = "html"
+    private let TEMP_FILE_NAME = "temp"
 
-    /// @private constructor
-    /// @see CreateFullscreenMessage method for construction
-    private init() {}
+    let fileManager = FileManager()
 
-    static func createFullscreenMessage(html: String, _ listener: FullscreenListenerInterface?, _ messageMonitor: MessageMonitor, _ isLocalImageUsed: Bool) -> FullScreenMessage? {
-        let newMessage: FullScreenMessage = FullScreenMessage()
-        newMessage.initMessage(html: html, listener, messageMonitor, false)
-        return newMessage
+    var isLocalImageUsed = false
+    var payload: String
+    var listener: FullscreenDelegate?
+    var webView: UIView?
+
+    private var messageService: MessageMonitorServicing {
+        return ServiceProvider.shared.messageMonitorService
     }
 
-    /// Initializes the platform message handler after the full screen message instance was created.
-    /// Note: we cannot do this step is done in the constructor, as the instance for FullscreenMessage is not created at that time
-    /// @param html - html content that will be loaded on message show
-    /// @param listener - listener which will be called on message show/dismiss/openUrl
-    /// @param messageMonnitor - MessageMonitor object which determines whether any display is already present
-    /// @param isLocalImageUsed If true, an image from the app bundle will be used for the fullscreen message.
-    func initMessage(html: String, _ listener: FullscreenListenerInterface?, _ messageMonitor: MessageMonitor, _ isLocalImageUsed: Bool) {
-        fullScreenMessageHandler = FullScreenUIHandler(payload: html, message: self, listener: listener, monitor: messageMonitor, isLocalImageUsed: isLocalImageUsed )
+    private var messagingDelegate: MessagingDelegate? {
+        return ServiceProvider.shared.messagingDelegate
     }
-}
 
-// MARK: - Protocol Methods
-extension FullScreenMessage: FullScreenMessageUiInterface {
+    /// Creates `FullscreenMessage` instance with the payload provided.
+    /// WARNING: This API consumes HTML/CSS/JS using an embedded browser control.
+    /// This means it is subject to all the risks of rendering untrusted web pages and running untrusted JS.
+    /// Treat all calls to this API with caution and make sure input is vetted for safety somewhere.
+    ///
+    /// - Parameters:
+    ///     - payload: String html content to be displayed with the message
+    ///     - listener: `FullscreenMessageDelegate` listener to listening the message lifecycle.
+    public convenience init(payload: String, listener: FullscreenDelegate?) {
+        self.init(payload: payload, listener: listener, isLocalImageUsed: false)
+    }
+
+    /// Creates `FullscreenMessage` instance with the payload provided.
+    /// WARNING: This API consumes HTML/CSS/JS using an embedded browser control.
+    /// This means it is subject to all the risks of rendering untrusted web pages and running untrusted JS.
+    /// Treat all calls to this API with caution and make sure input is vetted for safety somewhere.
+    ///
+    /// - Parameters:
+    ///     - payload: String html content to be displayed with the message
+    ///     - listener: `FullscreenMessageDelegate` listener to listening the message lifecycle.
+    ///     - isLocalImageUsed: If true, an image from the app bundle will be used for the fullscreen message.
+    public init(payload: String, listener: FullscreenDelegate?, isLocalImageUsed: Bool) {
+        self.payload = payload
+        self.listener = listener
+        self.isLocalImageUsed = isLocalImageUsed
+    }
 
     public func show() {
-        fullScreenMessageHandler?.show()
+        if messageService.show(message: self) ==  false {
+            return
+        }
+
+        DispatchQueue.main.async {
+            guard var newFrame: CGRect = self.calcFullscreenFrame() else {
+                Log.debug(label: self.LOG_PREFIX, "Failed to show the fullscreen message, newly created frame is nil.")
+                return
+            }
+            newFrame.origin.y = newFrame.size.height
+            if newFrame.size.width > 0.0 && newFrame.size.height > 0.0 {
+
+                let wkWebView = self.getConfiguredWebview(newFrame: newFrame)
+
+                // Fix for iPhone X to display content edge-to-edge
+                if #available(iOS 11, *) {
+                    wkWebView.scrollView.contentInsetAdjustmentBehavior = .never
+                }
+
+                // save the HTML payload to a local file if the cached image is being used
+                var useTempHTML = false
+                var cacheFolderURL: URL?
+                var tempHTMLFile: URL?
+                let cacheFolder: URL? = self.fileManager.getCacheDirectoryPath()
+                if cacheFolder != nil {
+                    cacheFolderURL = cacheFolder?.appendingPathComponent(self.DOWNLOAD_CACHE)
+                    tempHTMLFile = cacheFolderURL?.appendingPathComponent(self.TEMP_FILE_NAME).appendingPathExtension(self.HTML_EXTENSION)
+                    if !self.isLocalImageUsed {
+                        // We have to use loadFileURL so we can allow read access to these image files in the cache but loadFileURL
+                        // expects a file URL and not the string representation of the HTML payload. As a workaround, we can write the
+                        // payload string to a temporary HTML file located at cachePath/adbdownloadcache/temp.html and pass that file
+                        // URL to loadFileURL.
+                        do {
+                            try FileManager.default.createDirectory(atPath: cacheFolderURL?.path ?? "", withIntermediateDirectories: true, attributes: nil)
+                            try self.payload.write(toFile: tempHTMLFile?.path ?? "", atomically: true, encoding: .utf8)
+                            useTempHTML = true
+                        } catch {
+                            Log.debug(label: self.LOG_PREFIX, "Failed to save the temporary HTML file for fullscreen message \(error)")
+                        }
+                    }
+                }
+                // load the HTML string on WKWebview. If we are using the cached images, then use
+                // loadFileURL:allowingReadAccessToURL: to load the html from local file, which will give us the correct
+                // permission to read cached files
+                if useTempHTML {
+                    wkWebView.loadFileURL(URL.init(fileURLWithPath: tempHTMLFile?.path ?? ""), allowingReadAccessTo: URL.init(fileURLWithPath: cacheFolder?.path ?? ""))
+                } else {
+                    wkWebView.loadHTMLString(self.payload, baseURL: Bundle.main.bundleURL)
+                }
+
+                let keyWindow = UIApplication.shared.getKeyWindow()
+                keyWindow?.addSubview(wkWebView)
+                UIView.animate(withDuration: 0.3, animations: {
+                    var webViewFrame = wkWebView.frame
+                    webViewFrame.origin.y = 0
+                    wkWebView.frame = webViewFrame
+                }, completion: nil)
+
+                // Notifiying global listeners
+                self.listener?.onShow(message: self)
+                self.messagingDelegate?.onShow(message: self)
+            }
+        }
     }
 
-    public func openUrl(url: String) {
-        fullScreenMessageHandler?.openUrl(url: url)
+    public func dismiss() {
+        DispatchQueue.main.async {
+            if self.messageService.dismiss() ==  false {
+                return
+            }
+
+            self.dismissWithAnimation(animate: true)
+            // Notifiying all listeners
+            self.listener?.onDismiss(message: self)
+            self.messagingDelegate?.onDismiss(message: self)
+
+            // remove the temporary html if it exists
+            guard var cacheFolder: URL = self.fileManager.getCacheDirectoryPath() else {
+                return
+            }
+            cacheFolder.appendPathComponent(self.DOWNLOAD_CACHE)
+            cacheFolder.appendPathComponent(self.TEMP_FILE_NAME)
+            cacheFolder.appendPathComponent(self.HTML_EXTENSION)
+            let tempHTMLFilePath = cacheFolder.absoluteString
+
+            do {
+                try FileManager.default.removeItem(atPath: tempHTMLFilePath)
+            } catch {
+                Log.debug(label: self.LOG_PREFIX, "Unable to dismiss \(error)")
+            }
+        }
     }
 
-    public func remove() {
-        fullScreenMessageHandler?.dismiss()
+    // MARK: WKWebview delegate
+    public func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        if self.listener != nil {
+            guard let shouldOpenUrl = self.listener?.overrideUrlLoad(message: self, url: navigationAction.request.url?.absoluteString) else {
+                decisionHandler(.allow)
+                return
+            }
+            decisionHandler(shouldOpenUrl ? .allow : .cancel)
+
+        } else {
+            // if the API user doesn't provide any listner ( self.listener == nil ),
+            // set WKNavigationActionPolicyAllow as a default behaviour.
+            decisionHandler(.allow)
+        }
+    }
+
+    private func getConfiguredWebview(newFrame: CGRect) -> WKWebView {
+        let webViewConfiguration = WKWebViewConfiguration()
+
+        //Fix for media playback.
+        webViewConfiguration.allowsInlineMediaPlayback = true // Plays Media inline
+        webViewConfiguration.mediaTypesRequiringUserActionForPlayback = []
+        let wkWebView = WKWebView(frame: newFrame, configuration: webViewConfiguration)
+        wkWebView.navigationDelegate = self
+        wkWebView.scrollView.bounces = false
+        wkWebView.backgroundColor = UIColor.clear
+        wkWebView.isOpaque = false
+        wkWebView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        self.webView = wkWebView
+
+        return wkWebView
+    }
+
+    // MARK: web layout helpers
+    private func calcFullscreenFrame() -> CGRect? {
+        var newFrame = CGRect(x: 0, y: 0, width: 0, height: 0)
+        // x is always 0
+        newFrame.origin.x = 0
+        // for fullscreen, width and height are both full screen
+        let keyWindow = UIApplication.shared.getKeyWindow()
+        guard let screenBounds: CGSize = keyWindow?.frame.size else { return nil }
+        newFrame.size = screenBounds
+
+        // y is dependant on visibility and height
+        newFrame.origin.y = 0
+        return newFrame
+    }
+
+    private func dismissWithAnimation(animate: Bool) {
+        DispatchQueue.main.async {
+            UIView.animate(withDuration: animate ? 0.3: 0, animations: {
+                guard var newFrame: CGRect = self.calcFullscreenFrame() else {
+                    return
+                }
+                newFrame.origin.y = newFrame.size.height
+                self.webView?.frame = newFrame
+            }) { _ in
+                self.webView?.removeFromSuperview()
+                self.webView = nil
+            }
+        }
     }
 }
